requirements.txt
anyio==4.8.0
certifi==2025.1.31
et_xmlfile==2.0.0
fpdf==1.7.2
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
numpy==2.2.3
openpyxl==3.1.5
pandas==2.2.3
python-dateutil==2.9.0.post0
python-telegram-bot==21.10
pytz==2025.1
six==1.17.0
sniffio==1.3.1
tzdata==2025.1
